+++ sorting/C/utils.h
@@ -1,5 +1,5 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-void print_array(int *arr, size_t length);
+void print_array(int *arr, int length);
 #endif
